Refactored Redux items into store
Refactored items relating to Redux store into store folder.
Rename types.js to actionTypes.js.
