+++ front_end/src/store/reducers/auth.js
@@ -1,7 +1,7 @@
 import {
   USER_LOADED, USER_LOADING, AUTH_ERROR, LOGIN_SUCCESS, LOGIN_FAIL,
   LOGOUT_SUCCESS, REGISTER_SUCCESS, REGISTER_FAIL
-} from "../actions/types";
+} from "../actions/actionTypes";
 
 const initialState = {
   token: localStorage.getItem("token"),
