+++ front_end/src/store/reducers/index.js
