+++ front_end/src/store/reducers/messages.js
@@ -1,4 +1,4 @@
-import { CREATE_MESSAGE } from "../actions/types";
+import {CREATE_MESSAGE} from '../actions/actionTypes';
 
 const initialState = {};
